+++ src/query/delete.rs
@@ -9,15 +9,38 @@ use sea_query::DeleteStatement;
 #[derive(Clone, Debug)]
 pub struct Delete;
 
-/// Perform a delete operation on a model
+/// A request to delete an [`ActiveModel`](ActiveModelTrait).
+///
+/// The primary key must be set.
+/// Otherwise, it's impossible to generate the SQL condition and find the record.
+/// In that case, [`exec`][Self::exec] will return an error and not send any queries to the database.
+///
+/// If you want to use [`QueryTrait`] and access the generated SQL query,
+/// you need to convert into [`ValidatedDeleteOne`] first.
 #[derive(Clone, Debug)]
-pub struct DeleteOne<A>
-where
-    A: ActiveModelTrait,
-{
+pub struct DeleteOne<A: ActiveModelTrait>(pub(crate) Result<ValidatedDeleteOne<A>, DbErr>);
+
+/// A validated [`DeleteOne`] request, where the primary key is set
+/// and it's possible to generate the right SQL condition.
+#[derive(Clone, Debug)]
+pub struct ValidatedDeleteOne<A: ActiveModelTrait> {
     pub(crate) query: DeleteStatement,
-    pub(crate) error: Option<DbErr>,
     pub(crate) model: A,
+}
+
+impl<A: ActiveModelTrait> TryFrom<DeleteOne<A>> for ValidatedDeleteOne<A> {
+    type Error = DbErr;
+
+    fn try_from(value: DeleteOne<A>) -> Result<Self, Self::Error> {
+        value.0
+    }
+}
+
+impl<A: ActiveModelTrait> DeleteOne<A> {
+    /// Check whether the primary key is set and we can proceed with the operation.
+    pub fn validate(self) -> Result<ValidatedDeleteOne<A>, DbErr> {
+        self.try_into()
+    }
 }
 
 /// Perform a delete operation on multiple models
@@ -42,6 +65,8 @@ impl Delete {
     ///         id: 1,
     ///         name: "Apple Pie".to_owned(),
     ///     })
+    ///     .validate()
+    ///     .unwrap()
     ///     .build(DbBackend::Postgres)
     ///     .to_string(),
     ///     r#"DELETE FROM "cake" WHERE "cake"."id" = 1"#,
@@ -56,25 +81,43 @@ impl Delete {
     ///         id: ActiveValue::set(1),
     ///         name: ActiveValue::set("Apple Pie".to_owned()),
     ///     })
+    ///     .validate()
+    ///     .unwrap()
     ///     .build(DbBackend::Postgres)
     ///     .to_string(),
     ///     r#"DELETE FROM "cake" WHERE "cake"."id" = 1"#,
     /// );
     /// ```
+    //
+    // (non-doc comment for maintainers)
+    // Ideally, we would make this method fallible instead of stashing and delaying the error.
+    // But that's a bigger breaking change.
     pub fn one<E, A, M>(model: M) -> DeleteOne<A>
     where
         E: EntityTrait,
         A: ActiveModelTrait<Entity = E>,
         M: IntoActiveModel<A>,
     {
-        let myself = DeleteOne {
+        let mut myself = ValidatedDeleteOne {
             query: DeleteStatement::new()
                 .from_table(A::Entity::default().table_ref())
                 .to_owned(),
-            error: None,
             model: model.into_active_model(),
         };
-        myself.prepare()
+        // Build the SQL condition from the primary key columns.
+        for key in <A::Entity as EntityTrait>::PrimaryKey::iter() {
+            let col = key.into_column();
+            let av = myself.model.get(col);
+            match av {
+                ActiveValue::Set(value) | ActiveValue::Unchanged(value) => {
+                    myself = myself.filter(col.eq(value));
+                }
+                ActiveValue::NotSet => {
+                    return DeleteOne(Err(DbErr::PrimaryKeyNotSet { ctx: "DeleteOne" }));
+                }
+            }
+        }
+        DeleteOne(Ok(myself))
     }
 
     /// Delete many ActiveModel
@@ -103,28 +146,7 @@ impl Delete {
     }
 }
 
-impl<A> DeleteOne<A>
-where
-    A: ActiveModelTrait,
-{
-    pub(crate) fn prepare(mut self) -> Self {
-        for key in <A::Entity as EntityTrait>::PrimaryKey::iter() {
-            let col = key.into_column();
-            let av = self.model.get(col);
-            match av {
-                ActiveValue::Set(value) | ActiveValue::Unchanged(value) => {
-                    self = self.filter(col.eq(value));
-                }
-                ActiveValue::NotSet => {
-                    self.error = Some(DbErr::PrimaryKeyNotSet { ctx: "DeleteOne" });
-                }
-            }
-        }
-        self
-    }
-}
-
-impl<A> QueryFilter for DeleteOne<A>
+impl<A> QueryFilter for ValidatedDeleteOne<A>
 where
     A: ActiveModelTrait,
 {
@@ -146,7 +168,7 @@ where
     }
 }
 
-impl<A> QueryTrait for DeleteOne<A>
+impl<A> QueryTrait for ValidatedDeleteOne<A>
 where
     A: ActiveModelTrait,
 {
@@ -187,7 +209,7 @@ where
 #[cfg(test)]
 mod tests {
     use crate::tests_cfg::{cake, fruit};
-    use crate::{DbBackend, entity::*, query::*};
+    use crate::{DbBackend, entity::*, query::*, *};
 
     #[test]
     fn delete_1() {
@@ -196,6 +218,8 @@ mod tests {
                 id: 1,
                 name: "Apple Pie".to_owned(),
             })
+            .validate()
+            .unwrap()
             .build(DbBackend::Postgres)
             .to_string(),
             r#"DELETE FROM "cake" WHERE "cake"."id" = 1"#,
@@ -205,6 +229,8 @@ mod tests {
                 id: ActiveValue::set(1),
                 name: ActiveValue::set("Apple Pie".to_owned()),
             })
+            .validate()
+            .unwrap()
             .build(DbBackend::Postgres)
             .to_string(),
             r#"DELETE FROM "cake" WHERE "cake"."id" = 1"#,